✨ ZA0BEE code: decompose route decision
