+++ .agentplane/tasks/202605281934-ZA0BEE/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605281934-ZA0BEE/route-decision-module-decomposition",
+  "created_at": "2026-05-28T19:34:55.525Z",
+  "diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "last_verified_at": "2026-05-28T19:39:40.365Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605281934-ZA0BEE",
+  "updated_at": "2026-05-28T19:34:55.525Z",
+  "verify": {
+    "status": "pass"
+  }
+}
