+++ .agentplane/tasks/202605281934-ZA0BEE/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T19:34:55.525Z
+
+## Task
+
+- Task: `202605281934-ZA0BEE`
+- Title: Route decision module decomposition
+- Status: DOING
+- Branch: `task/202605281934-ZA0BEE/route-decision-module-decomposition`
+- Canonical task record: `.agentplane/tasks/202605281934-ZA0BEE/README.md`
+
+## Verification
+
+- State: ok
+- Note: Route decision decomposition verified: facade reduced to 335 lines, blockers/next-action/types extracted into focused modules, route-decision CLI tests passed, typecheck passed, hotspot threshold check passed, format:changed passed.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T19:34:55.525Z
+- Branch: task/202605281934-ZA0BEE/route-decision-module-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
